requirements.txt
ydb
requests
black
pydantic
alice-types
pytest
pytest-asyncio
pytest-dotenv
pytest-aiohttp